% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire.R
\name{acquire_ct}
\alias{acquire_ct}
\title{Download ONS Consumer Trends dataset}
\usage{
acquire_ct(path)
}
\arguments{
\item{path}{Folder to put the downloaded data in. If missing a tempfile will be created.}
}
\value{
The file path and name of the downloaded file.
}
\description{
Download ONS Consumer Trends dataset
}
