% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{check_indicator}
\alias{check_indicator}
\title{Test whether a string is a valid Pocketbook indicator id}
\usage{
check_indicator(string)
}
\arguments{
\item{string}{String containing a Pocketbook indicator id}
}
\value{
A logical
}
\description{
A Pocketbook indicator can be uniquely identified from 2 hierarchical
components separated by periods:
\itemize{
\item section (1-9)
\item indicator (1-99)
}
}
\examples{
# A valid id
check_indicator("2.3")

# An invalid id
check_indicator("10.3")
}
\seealso{
\code{\link[=parse_indicator]{parse_indicator()}} extracts the components of an indicator id

Other Helpers: 
\code{\link{parse_indicator}()},
\code{\link{s3_bucket}()},
\code{\link{save_csv}()},
\code{\link{save_graphic}()}
}
\concept{Helpers}
