% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsp-package.R
\docType{package}
\name{fsp-package}
\alias{fsp}
\alias{fsp-package}
\title{fsp: Utilities for the Food Statistics Pocketbook}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://foodchainstats.github.io/fsp/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
