% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag-workforce.R
\name{get_ag_workforce}
\alias{get_ag_workforce}
\title{Extract data from an Agricultural workforce ods file}
\usage{
get_ag_workforce(file, sheet = "Ag_workforce_by_country")
}
\arguments{
\item{file}{A spreadsheet in ods format. If omitted, will be downloaded using
\code{\link[=acquire_ag_workforce]{acquire_ag_workforce()}}.}

\item{sheet}{Spreadsheet tab name or number}
}
\value{
A tibble of workforce data
}
\description{
Extract data from an Agricultural workforce ods file
}
\examples{
\dontrun{
# download and extract in one go
ag <- get_ag_workforce()

# or download separately and extract 
ag <- acquire_ag_workforce()
data <- get_ag_workforce(ag)

}
}
\seealso{
Other {Agricultural workforce}: 
\code{\link{acquire_ag_workforce}()},
\code{\link{compile_ag_workforce}()},
\code{\link{url_ag_workforce}()}
}
\concept{{Agricultural workforce}}
