% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{get_ct_quarter}
\alias{get_ct_quarter}
\title{Get Consumer Trends quarterly data}
\usage{
get_ct_quarter(rawfile)
}
\arguments{
\item{rawfile}{An ONS ct.csv file}
}
\value{
A tibble of quarterly series data
}
\description{
Get Consumer Trends quarterly data
}
