% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{get_ct_year}
\alias{get_ct_year}
\title{Get Consumer Trends annual data}
\usage{
get_ct_year(rawfile)
}
\arguments{
\item{rawfile}{An ONS ct.csv file}
}
\value{
A tibble of annual series data
}
\description{
Get Consumer Trends annual data
}
