% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unctad.R
\name{get_unctad_metadata}
\alias{get_unctad_metadata}
\title{Get UNCTAD metadata}
\usage{
get_unctad_metadata(file)
}
\arguments{
\item{file}{An UNCTAD download file (see \code{\link[=acquire_unctad]{acquire_unctad()}}). If omitted data
will be downloaded automatically}
}
\value{
A tibble of UNCTAD commodity price metadata
}
\description{
Get UNCTAD metadata
}
\examples{
\dontrun{
metadata <- get_unctad_metadata()

acquire_unctad("~/downloads")

get_unctad_metadata("~/downloads/commodityprices.csv")  
}
}
\seealso{
Other UNCTAD: 
\code{\link{acquire_unctad}()},
\code{\link{get_unctad}()}
}
\concept{UNCTAD}
