% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-jobs03.R
\name{jobs03_AUK_sectors}
\alias{jobs03_AUK_sectors}
\title{Dataset of  AUK JOBS03 SIC codes and food chain sectors}
\usage{
jobs03_AUK_sectors()
}
\value{
A tibble of SIC codes and sectors.
}
\description{
The 'official' SIC definition of the food chain that Defra uses is described
in the \href{https://www.gov.uk/government/statistics/food-statistics-pocketbook/food-statistics-in-your-pocket#glossary}{pocketbook glossary}.
This dataset uses the SIC codes included in JOBS03. It is useful to join
against raw JOBS03 data to extract food sectors of innterest.
}
\seealso{
Other {JOBS03}: 
\code{\link{acquire_jobs03}()},
\code{\link{get_jobs03}()},
\code{\link{jobs03_sectors}()},
\code{\link{url_jobs03}()}
}
\concept{{JOBS03}}
