% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_csv}
\alias{save_csv}
\title{Save Pocketbook data for publication to the S3 bucket}
\usage{
save_csv(data, indicator_id, indicator_desc = "")
}
\arguments{
\item{data}{A data frame}

\item{indicator_id}{A valid Pocketbook indicator id. See \code{\link[=check_indicator]{check_indicator()}}.}

\item{indicator_desc}{A title for the graphic}
}
\value{
Saves a png and svg file to the S3 bucket. The filename will be the
indicator id, description and current date (with all punctuation removed),
delimited by underscores '_'.
}
\description{
Data is saved as a csv into a folder structure on the bucket. This is
currently 'section_x/iS_I/output/graphics'. Filenames are derived from the
indicator id, description and current date. The bucket location is encoded in
\code{\link[=s3_bucket]{s3_bucket()}}.
}
\examples{
\dontrun{
save_csv(mtcars, "1.1", "mtcars data")
}
}
\seealso{
\code{\link[=save_graphic]{save_graphic()}} for saving images
}
