% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_graphic}
\alias{save_graphic}
\title{Save a Pocketbook graphic to the S3 bucket}
\usage{
save_graphic(graphic, indicator_id, indicator_desc = "")
}
\arguments{
\item{graphic}{A \code{ggplot} chart object}

\item{indicator_id}{A valid Pocketbook indicator id. See \code{\link[=check_indicator]{check_indicator()}}.}

\item{indicator_desc}{A title for the graphic}
}
\value{
Saves a png and svg file to the S3 bucket. The filename will be the
indicator id, description and current date (with all punctuation removed),
delimited by underscores '_'.
}
\description{
Graphics are saved into a folder structure on the bucket. This is currently
'section_x/iS_I/output/graphics'. Filenames are derived from the indicator
id, description and curent date. The bucket location is encoded in
\code{\link[=s3_bucket]{s3_bucket()}}.
}
\examples{
\dontrun{
chart <- ggplot2::ggplot(mtcars) + 
ggplot2::geom_point(ggplot2::aes(x = mpg, y = wt))

save_graphic(chart, "1.1", "Test graphic")
#file created will be '1_1_test_graphic.png/svg'
}
}
\seealso{
\code{\link[=save_csv]{save_csv()}} for saving data

Other Helpers: 
\code{\link{check_indicator}()},
\code{\link{parse_indicator}()},
\code{\link{s3_bucket}()},
\code{\link{save_csv}()}
}
\concept{Helpers}
