% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{url_abs}
\alias{url_abs}
\title{URL for the current Annual Business Survey dataset}
\usage{
url_abs()
}
\value{
The url for the dataset
}
\description{
Returns the latest file from the \href{https://www.ons.gov.uk/businessindustryandtrade/business/businessservices/datasets/uknonfinancialbusinesseconomyannualbusinesssurveysectionsas}{ONS}
}
\examples{
url_abs()
}
