% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{url_bpe}
\alias{url_bpe}
\title{URL for Business Population Estimates}
\usage{
url_bpe(year = 2024)
}
\arguments{
\item{year}{}
}
\value{
The url for the BPE collection page
}
\description{
Returns the collection page for \href{https://www.gov.uk/government/collections/business-population-estimates}{Business Population Estimates}
}
\examples{
url_bpe()
}
