% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{url_ct}
\alias{url_ct}
\title{URL for the current Consumer Trends dataset}
\usage{
url_ct()
}
\value{
The url for the dataset.
}
\description{
URL for the current Consumer Trends dataset
}
\examples{
url_ct()
}
\seealso{
Other {Consumer Trends}: 
\code{\link{acquire_ct}()},
\code{\link{get_ct_metadata}()},
\code{\link{get_ct_quarter}()},
\code{\link{get_ct_year}()}
}
\concept{{Consumer Trends}}
