% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{url_jobs03}
\alias{url_jobs03}
\title{URL for the current JOBS03 dataset}
\usage{
url_jobs03()
}
\value{
The url for the dataset.
}
\description{
Scrapes the filename of the latest dataset from
\href{https://www.ons.gov.uk/employmentandlabourmarket/peopleinwork/employmentandemployeetypes/datasets/employeejobsbyindustryjobs03}{JOBS03}
}
\examples{
\dontrun{
url_jobs03()
}
}
\seealso{
Other {JOBS03}: 
\code{\link{acquire_jobs03}()},
\code{\link{get_jobs03}()},
\code{\link{jobs03_AUK_sectors}()},
\code{\link{jobs03_sectors}()}
}
\concept{{JOBS03}}
