% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{url_unctad}
\alias{url_unctad}
\title{URL for UNCTAD commodity price indices data}
\usage{
url_unctad()
}
\value{
the url for commodity price indices
}
\description{
URL for UNCTAD commodity price indices data
}
\examples{
\dontrun{
url_unctad()
}
}
