% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire.R
\name{acquire_ag_workforce}
\alias{acquire_ag_workforce}
\title{Download Defra Agricultural Workforce dataset}
\usage{
acquire_ag_workforce(path)
}
\arguments{
\item{path}{Folder to put the downloaded data in. If missing a tempfile will
be created. If specified the downloaded file will be named
'ag_workforce.ods'.}
}
\value{
The file path and name of the downloaded file.
}
\description{
The downloaded file will be an ODS spreadsheet.
}
\examples{
\dontrun{
jobs03 <- acquire_ag_workforce()

file <- acquire_ag_workforce("~/downloads")
}
}
\seealso{
Other Agricultural workforce: 
\code{\link{compile_ag_workforce}()},
\code{\link{get_ag_workforce}()}
}
\concept{Agricultural workforce}
