% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire.R
\name{acquire_unctad}
\alias{acquire_unctad}
\title{Extract UNCTAD Commodity Price indices}
\usage{
acquire_unctad(path)
}
\arguments{
\item{path}{Folder to put the downloaded data in. If missing a tempfile will
be created. If specified the downloaded file will be named
'commodityprices.csv'.}
}
\value{
The file path and name of the downloaded file.
}
\description{
Downloads and unzips UNCTAD data (see \code{\link[=url_unctad]{url_unctad()}}). Internally uses the
\code{archive} library which requires libarchive-dev to be installed on Ubuntu
machines.
}
\examples{
\dontrun{
a <- acquire_unctad(a)
a

acquire_unctad("~/downloads")
}
}
\seealso{
Other UNCTAD: 
\code{\link{get_unctad}()},
\code{\link{get_unctad_metadata}()}
}
\concept{UNCTAD}
