% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire.R
\name{acquire_unctad}
\alias{acquire_unctad}
\title{Download UNCTAD Commodity Price indices}
\usage{
acquire_unctad(path)
}
\arguments{
\item{path}{Folder to put the downloaded data in. If missing a tempfile will
be created. If specified the downloaded file will be named
'commodityprices.csv'.}
}
\value{
The file path and name of the downloaded file.
}
\description{
Download UNCTAD Commodity Price indices
}
\examples{
\dontrun{
unctad <- acquire_unctad()

file <- acquire_unctad("~/downloads")
}
}
