% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-ag.R
\name{compile_ag_workforce}
\alias{compile_ag_workforce}
\title{Compile GB Agricultural workforce data}
\usage{
compile_ag_workforce()
}
\value{
A dataframe of quarterly data
}
\description{
Compile GB Agricultural workforce data
}
\examples{
\dontrun{
data <- compile_ag_workforce()
}
}
\seealso{
Other {Agricultural workforce}: 
\code{\link{acquire_ag_workforce}()},
\code{\link{get_ag_workforce}()},
\code{\link{url_ag_workforce}()}
}
\concept{{Agricultural workforce}}
