% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsp-package.R
\docType{package}
\name{fsp-package}
\alias{fsp}
\alias{fsp-package}
\title{fsp: Utilities for the Food Statistics Pocketbook}
\description{
Experimental functions to assist with producing the Pocketbook.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://foodchainstats.github.io/fsp/}
}

}
\author{
\strong{Maintainer}: David Lee \email{david.lee@defra.gov.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
