% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag-workforce.R
\name{get_ag_workforce}
\alias{get_ag_workforce}
\title{Extract data from an Agricultural workforce ods file}
\usage{
get_ag_workforce(file)
}
\arguments{
\item{file}{A spreadsheet in ods format. If omitted, will be downloaded.}
}
\value{
A tibble of workforce data
}
\description{
Extract data from an Agricultural workforce ods file
}
