% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{get_ct_quarter}
\alias{get_ct_quarter}
\title{Get Consumer Trends quarterly data}
\usage{
get_ct_quarter(rawfile)
}
\arguments{
\item{rawfile}{An ONS ct.csv file}
}
\value{
A tibble of quarterly series data
}
\description{
Get Consumer Trends quarterly data
}
\examples{
\dontrun{
get_ct_quarter()
}
}
\seealso{
Other {Consumer Trends}: 
\code{\link{acquire_ct}()},
\code{\link{get_ct_metadata}()},
\code{\link{get_ct_year}()},
\code{\link{url_ct}()}
}
\concept{{Consumer Trends}}
