% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{get_ct_year}
\alias{get_ct_year}
\title{Get Consumer Trends annual data}
\usage{
get_ct_year(rawfile)
}
\arguments{
\item{rawfile}{An ONS ct.csv file}
}
\value{
A tibble of annual series data
}
\description{
Get Consumer Trends annual data
}
\examples{
\dontrun{
egt_ct_year()
}
}
\seealso{
Other {Consumer Trends}: 
\code{\link{get_ct_metadata}()},
\code{\link{get_ct_quarter}()}
}
\concept{{Consumer Trends}}
