% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unctad.R
\name{get_unctad}
\alias{get_unctad}
\title{Get UNCTAD data}
\usage{
get_unctad(file)
}
\arguments{
\item{file}{An UNCTAD bulk download csv}
}
\value{
A tibble of UNCTAD commodity price data
}
\description{
Get UNCTAD data
}
\examples{
\dontrun{
get_unctad()
}
}
