% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unctad.R
\name{get_unctad_metadata}
\alias{get_unctad_metadata}
\title{Get UNCTAD metadata}
\usage{
get_unctad_metadata()
}
\value{
A tibble of UNCTAD commodity price metadata
}
\description{
Get UNCTAD metadata
}
\examples{
\dontrun{
get_unctad_metadata()
}
}
