% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unctad.R
\name{get_unctad_metadata}
\alias{get_unctad_metadata}
\title{Get UNCTAD metadata}
\usage{
get_unctad_metadata(file)
}
\arguments{
\item{file}{An UNCTAD 7zip download file (see \code{\link[=url_unctad]{url_unctad()}} for where to get
it)}
}
\value{
A tibble of UNCTAD commodity price metadata
}
\description{
Get UNCTAD metadata
}
\examples{
a <- system.file("extdata", "unctad-example.csv.7z", package = "fsp")
a
get_unctad_metadata(a)
}
\seealso{
Other {UNCTAD}: 
\code{\link{acquire_unctad}()},
\code{\link{get_unctad}()},
\code{\link{url_unctad}()}
}
\concept{{UNCTAD}}
