% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-jobs03.R
\name{jobs03_AUK_sectors}
\alias{jobs03_AUK_sectors}
\title{Dataset of  AUK JOBS03 SIC codes and food chain sectors}
\usage{
jobs03_AUK_sectors()
}
\value{
A tibble of SIC codes and sectors.
}
\description{
Deprecated. See the parameters of \code{\link[=jobs03_sectors]{jobs03_sectors()}} for ways of excluding
fishing or total industry definitions.
}
\examples{
jobs03_AUK_sectors()
}
\seealso{
Other {JOBS03}: 
\code{\link{acquire_jobs03}()},
\code{\link{get_jobs03}()},
\code{\link{jobs03_sectors}()},
\code{\link{url_jobs03}()}
}
\concept{{JOBS03}}
