% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-jobs03.R
\name{jobs03_sectors}
\alias{jobs03_sectors}
\title{Dataset of JOBS03 SIC codes and food chain sectors}
\usage{
jobs03_sectors()
}
\value{
A tibble of SIC codes and sectors
}
\description{
Dataset of JOBS03 SIC codes and food chain sectors
}
\examples{
jobs03_sectors()
}
