% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-jobs03.R
\name{jobs03_sectors}
\alias{jobs03_sectors}
\title{Dataset of JOBS03 SIC codes and food chain sectors}
\usage{
jobs03_sectors(include_fishing = TRUE, include_total = TRUE)
}
\arguments{
\item{include_fishing}{logical: include the definition of the fishing industry}

\item{include_total}{logical: include the definition of total UK industry}
}
\value{
A tibble of SIC codes and sectors.
}
\description{
The 'official' SIC definition of the food chain that Defra uses is described
in the \href{https://www.gov.uk/government/statistics/food-statistics-pocketbook/food-statistics-in-your-pocket#glossary}{pocketbook glossary}.
This dataset uses the SIC codes included in JOBS03. It is useful to join
against raw JOBS03 data to extract food sectors of interest.
}
\examples{
jobs03_sectors()
}
\seealso{
Other JOBS03: 
\code{\link{acquire_jobs03}()},
\code{\link{get_jobs03}()},
\code{\link{jobs03_AUK_sectors}()}
}
\concept{JOBS03}
