% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{url_ag_workforce}
\alias{url_ag_workforce}
\title{URL for Defra Agricultural Woorkforce data}
\usage{
url_ag_workforce()
}
\value{
The landing page that contains the dataset. get_jobs03_data() uses
this function and some webscraping to download and extract thte actual data
file.
}
\description{
URL for Defra Agricultural Woorkforce data
}
