% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{url_ag_workforce}
\alias{url_ag_workforce}
\title{URL for Defra Agricultural Workforce data}
\usage{
url_ag_workforce()
}
\value{
The url for the dataset.
}
\description{
Scrapes the filename of the latest data from the Defra \href{https://www.gov.uk/government/statistical-data-sets/agricultural-workforce-in-the-united-kingdom-at-1-june}{Agricultural workforce}
site. The file is an ODS spreadsheet.
}
\examples{
\dontrun{
url_ag_workforce()
}
}
