% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{url_ct}
\alias{url_ct}
\title{URL for the current Consumer Trends dataset}
\usage{
url_ct()
}
\value{
The url for the dataset.
}
\description{
URL for the current Consumer Trends dataset
}
\examples{
url_ct()
}
