% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{url_unctad}
\alias{url_unctad}
\title{URL for UNCTAD commodity price indices data}
\usage{
url_unctad()
}
\value{
the url for commodity price indices
}
\description{
Scrapes the filename of the latest data from the UNCTAD \href{https://unctadstat.unctad.org/EN/BulkDownload.html}{bulk download} page. The file
is a 7zip archive file.
}
\examples{
\dontrun{
url_unctad()
}
}
