% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{url_unctad}
\alias{url_unctad}
\title{URL for UNCTAD commodity price indices data}
\usage{
url_unctad()
}
\value{
the url for commodity price indices
}
\description{
Currently the data cannot be scraped (or I am not clever enough to do so),
and so this function simply returns the url where it can be manually
downloaded. The file is a 7zip archive. The \code{\link[=acquire_unctad]{acquire_unctad()}}, \code{\link[=get_unctad]{get_unctad()}}
and \code{\link[=get_unctad_metadata]{get_unctad_metadata()}} functions all require access to a downloaded
7zip. This package includes an example file - access it using:
\code{system.file("extdata", "unctad-example.csv.7z", package = "fsp")}
}
\examples{
\dontrun{
url_unctad()
}
}
\concept{{UNCTAD}}
